Add files via upload
Program now runs in a fully threaded manner and both players and decks write to files. Unless two players win simultaneously, only one winner is permitted. Two players winning simultaneously seems oddly common but that might just be a result of the packs provided. This iteration of the program seems to be immune to deadlocking. The one thing I don't understand is why the decks end up with varying amounts of cards at the end of the game. It doesn't seem to cause any run time problems but it doesn't seem like it should be happening this way.