Add files via upload
Only change to functionality: Added a check to make sure a pack file was not larger than 8*(number of players), as that seemed to be the last thing that was leading to errors. Otherwise, removed most commented-out testing code and added more useful comments to some uncommented methods and classes to improve readability.